+++ vr_to_joystick/mappings/test_mapping.py
@@ -4,12 +4,12 @@ import openvr
 from pyvjoy.vjoydevice import HID_USAGE_X, HID_USAGE_Y, HID_USAGE_Z, HID_USAGE_RX, HID_USAGE_RY, HID_USAGE_RZ, HID_USAGE_SL0
 
 from vr_to_joystick.controller_mapping import ControllerMapping
-from vr_to_joystick.mappings.nodes.axis import ControllerAxis, VXAxis, VYAxis, VZAxis, XAxis, YAxis, ZAxis
-from vr_to_joystick.mappings.nodes.axis_helpers import ScaleAxis
-from vr_to_joystick.mappings.nodes.button import DirectButton
-from vr_to_joystick.mappings.nodes.button_helpers import AxisThresholdButton
-from vr_to_joystick.mappings.nodes.types import Axis, Button
-from vr_to_joystick.mappings.nodes.vr_system_state import ControllerStateByType, ControllerStateGenerator, VrSystemState
+from vr_to_joystick.nodes.axis import ControllerAxis, VXAxis, VYAxis, VZAxis, XAxis, YAxis, ZAxis
+from vr_to_joystick.nodes.axis_helpers import ScaleAxis
+from vr_to_joystick.nodes.button import DirectButton
+from vr_to_joystick.nodes.button_helpers import AxisThresholdButton
+from vr_to_joystick.nodes.types import Axis, Button
+from vr_to_joystick.nodes.vr_system_state import ControllerStateByType, ControllerStateGenerator, VrSystemState
 
 ATAN_AXIS_SCALAR = 1 / (2 * math.pi)
 TOUCHPAD_EDGE_BUTTON_THRESHOLD = 0.8
