+++ vr_to_joystick/nodes/axis.py
@@ -1,7 +1,7 @@
 import math
 from typing import Any, Hashable, Literal
 
-from vr_to_joystick.mappings.nodes.vr_system_state import ControllerStateConsumer
+from vr_to_joystick.nodes.vr_system_state import ControllerStateConsumer
 
 
 DirectAxis = ControllerStateConsumer[float]
