+++ vr_to_joystick/nodes/value_generator.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 from abc import abstractmethod
 from typing import Any, Hashable, TypeVar, Generic
 
-from vr_to_joystick.mappings.multiton import MultitonNode
+from vr_to_joystick.nodes.multiton import MultitonNode
 
 O = TypeVar('O')
