+++ vr_to_joystick/nodes/vr_system_state.py
@@ -3,7 +3,7 @@ from typing import Any, Callable, Hashable, Iterable, Literal, TypeVar, TypedDic
 
 import openvr
 
-from vr_to_joystick.mappings.nodes.value_generator import ValueGenerator
+from vr_to_joystick.nodes.value_generator import ValueGenerator
 
 DeviceClass = Literal['controller', 'generic_tracker', 'hmd']
