+++ vr_to_joystick/nodes/wheel.py
@@ -1,8 +1,8 @@
 import math
 from typing import Any
 
-from vr_to_joystick.mappings.nodes.types import Axis
-from vr_to_joystick.mappings.nodes.vr_system_state import ControllerStateGenerator
+from vr_to_joystick.nodes.types import Axis
+from vr_to_joystick.nodes.vr_system_state import ControllerStateGenerator
 
 
 # Axis that reports the angle between a left and right controller.
